+++ crates/atuin-client/src/import/mod.rs
@@ -103,7 +103,10 @@ fn is_file(p: PathBuf) -> Result<PathBuf> {
     if p.is_file() {
         Ok(p)
     } else {
-        bail!("Could not find history file {:?}. Try setting $HISTFILE", p)
+        bail!(
+            "Could not find history file {:?}. Try setting and exporting $HISTFILE",
+            p
+        )
     }
 }
 fn is_dir(p: PathBuf) -> Result<PathBuf> {
@@ -111,7 +114,7 @@ fn is_dir(p: PathBuf) -> Result<PathBuf> {
         Ok(p)
     } else {
         bail!(
-            "Could not find history directory {:?}. Try setting $HISTFILE",
+            "Could not find history directory {:?}. Try setting and exporting $HISTFILE",
             p
         )
     }