+++ crates/atuin-daemon/src/server.rs
@@ -4,6 +4,7 @@ use atuin_client::encryption;
 use atuin_client::history::store::HistoryStore;
 use atuin_client::record::sqlite_store::SqliteStore;
 use atuin_client::settings::Settings;
+#[cfg(unix)]
 use std::path::PathBuf;
 use std::sync::Arc;
 use time::OffsetDateTime;
@@ -227,7 +228,7 @@ async fn start_server(settings: Settings, history: HistoryService) -> Result<()>
     use tokio_stream::wrappers::TcpListenerStream;
 
     let port = settings.daemon.tcp_port;
-    let url = format!("127.0.0.1:{}", port);
+    let url = format!("127.0.0.1:{port}");
     let tcp = TcpListener::bind(url).await?;
     let tcp_stream = TcpListenerStream::new(tcp);